+++ src/main/kotlin/br/com/cooperativa/votacao/domain/persist/VotePersist.kt
@@ -3,11 +3,14 @@ package br.com.cooperativa.votacao.domain.persist
 import br.com.cooperativa.votacao.domain.dto.VoteType
 import br.com.cooperativa.votacao.util.cleanCodeText
 import br.com.cooperativa.votacao.util.createId
+import br.com.cooperativa.votacao.util.now
 import org.springframework.data.annotation.Id
+import java.time.LocalDateTime
 
 class VotePersist(
     id: String = createId(), // CPF
-    val vote: VoteType = VoteType.NOT_SELECTED
+    val vote: VoteType = VoteType.NOT_SELECTED,
+    val createdAt: LocalDateTime = now()
 ) {
 
     @Id
@@ -22,11 +25,13 @@ class VotePersist(
     companion object {
         fun build(
             id: String,
-            vote: VoteType
+            vote: VoteType,
+            createdAt: LocalDateTime = now()
         ): VotePersist {
             return VotePersist(
                 id = id,
-                vote = vote
+                vote = vote,
+                createdAt = createdAt
             )
         }
     }