+++ src/main/kotlin/br/com/cooperativa/votacao/util/DateUtils.kt
@@ -1,5 +1,6 @@
 package br.com.cooperativa.votacao.util
 
+import java.time.Duration
 import java.time.Instant
 import java.time.LocalDateTime
 import java.time.ZoneId
@@ -14,6 +15,11 @@ fun getDefaultZoneId(): ZoneId = ZoneId.of("America/Sao_Paulo")
 
 fun zonedNow(): ZonedDateTime = ZonedDateTime.now(getDefaultZoneId())
 
+fun now(): LocalDateTime = zonedNow().toLocalDateTime()
+
+fun LocalDateTime.duration(other: LocalDateTime) = Duration.between(this, other)
+
+fun ZonedDateTime.duration(other: ZonedDateTime) = Duration.between(this, other)
 
 fun LocalDateTime.toZonedDateTime(zoneId: ZoneId = getDefaultZoneId()): ZonedDateTime = this.atZone(zoneId)