+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### 2024-10-31
+
+* Criação das classes DTO: AgendaDTO, VoteDTO, SummaryAgendaDTO;
+* Correção das classes AgendaPersist e VotePersist, utilizando LocalDateTime;
+* Ajustes e criação de classes de testes unitários;
+* Adicionado script compile e compile.bat
+* 
+
 ### 2024-10-30
 
 * Criação do projeto;